+++ source/gui/interface_definitons.cpp
@@ -275,6 +275,7 @@ namespace zlInterface {
         cMap1Idx = static_cast<size_t>(loadPara(zlState::colourMap1Idx::ID));
         cMap2Idx = static_cast<size_t>(loadPara(zlState::colourMap2Idx::ID));
         fftOrderIdx = static_cast<int>(loadPara(zlState::ffTOrder::ID));
+        dynLink.store(static_cast<bool>(loadPara(zlState::dynLink::ID)));
     }
 
     void UIBase::saveToAPVTS() const {
@@ -322,5 +323,6 @@ namespace zlInterface {
         savePara(zlState::colourMap1Idx::ID, zlState::colourMapIdx::convertTo01(static_cast<int>(cMap1Idx)));
         savePara(zlState::colourMap2Idx::ID, zlState::colourMapIdx::convertTo01(static_cast<int>(cMap2Idx)));
         savePara(zlState::ffTOrder::ID, zlState::ffTOrder::convertTo01(fftOrderIdx));
+        savePara(zlState::dynLink::ID, zlState::dynLink::convertTo01(dynLink.load()));
     }
 }